# Improve startup performance by lazy-compiling mashumaro serialization…